+++ locomotion-sdk/src/Components/PageHeader/index.js
@@ -9,7 +9,7 @@ import i18n from '../../I18n';
 const HamburgerIconSource = require('../../assets/menu.png');
 
 const PageHeader = ({
-  title, icon, onIconPress, iconSide, displayIcon, width, height, children,showSkipButton, onPressSkip,
+  title, icon, onIconPress, iconSide, displayIcon, width, height, children, showSkipButton, onPressSkip,
 }) => (
   <Header>
     {displayIcon !== false
