+++ locomotion-sdk/src/context/state/index.js
@@ -7,3 +7,17 @@ export const StateProvider = ({ reducer, initialState, children }) => (
   </StateContext.Provider>
 );
 export const useStateValue = () => useContext(StateContext);
+
+const getPopupKey = key => `popup_${key}`;
+
+export const getTogglePopupsState = () => {
+  const [state, dispatch] = useContext(StateContext);
+
+  return [id => state && (state[getPopupKey(id)] || false), (id, open) => dispatch({
+    type: 'changeState',
+    payload: {
+      [getPopupKey('ridePopupsStatus')]: open,
+      [getPopupKey(id)]: open,
+    },
+  })];
+};
