+++ locomotion-sdk/src/pages/AuthScreens/Onboarding/Avatar.js
@@ -9,7 +9,7 @@ import {
 import Header from './Header';
 import ScreenText from './ScreenText/index';
 import ThumbnailPicker from '../../../Components/ThumbnailPicker';
-import { ONBOARDING_PAGE_NAMES } from '../../../pages/consts';
+import { ONBOARDING_PAGE_NAMES } from '../../routes';
 
 const Avatar = () => {
   const {
