+++ locomotion-sdk/src/pages/AuthScreens/Onboarding/Email.js
@@ -7,7 +7,7 @@ import { ErrorText, PageContainer, SafeView } from './styles';
 import i18n from '../../../I18n';
 import Header from './Header';
 import ScreenText from './ScreenText';
-import { ONBOARDING_PAGE_NAMES } from '../../../pages/consts';
+import { ONBOARDING_PAGE_NAMES } from '../../routes';
 
 
 const Email = () => {
