+++ locomotion-sdk/src/pages/AuthScreens/Onboarding/Header.js
@@ -16,7 +16,7 @@ const Header = ({ title, page }) => {
   };
 
   const goBack = () => {
-    navigation.goBack()
+    navigation.goBack();
   };
   return (
     <PageHeader
